Add post code, mobile validation and auto logout.

+++ ParkingHere-web-client/src/appModules/account/store/UserLoginStore.ts
@@ -151,11 +151,13 @@ export const useUserLoginStore = defineStore({
     },
 
     logOutUserWhenTokenExpired() {
-      if (moment(this.getTokenExpiration).diff(moment(), 'minutes') < 0) {
-        service.removeUserFromLocalStorage()
-        this.loggedIn = false
-        toast.success(i18n.global.t('TokenExpired'))
-        this.router.push('/')
+      if (this.loggedIn) {
+        if (moment(this.getTokenExpiration).diff(moment(), 'minutes') < 0) {
+          service.removeUserFromLocalStorage()
+          this.loggedIn = false
+          toast.success(i18n.global.t('TokenExpired'))
+          this.router.push('/')
+        }
       }
     }
   }
